+++ mail_base/__init__.py
@@ -1,3 +1,4 @@
+# License LGPL-3.0 (https://www.gnu.org/licenses/lgpl.html)
 
 from . import models
 from . import controllers