+++ mail_to/doc/changelog.rst
@@ -1,3 +1,9 @@
+`1.1.0`
+-------
+
+- **New**: channels are displayed in recipients
+
+
 `1.0.1`
 -------